Added left border to description
