+++ src/UI-Kit/index.tsx
@@ -1,9 +1,17 @@
 import Button from './Button';
+import CaseStudyTile from './CaseStudyTile';
 import Footer from './Footer';
 import Header from './Header';
-import CaseStudyTile from './CaseStudyTile';
 import Link from './Link';
 import Navbar from './Navbar';
 import Tile from './Tile';
 
-export { Button, Link, CaseStudyTile, Navbar, Tile, Header, Footer };
+export {
+  Button,
+  CaseStudyTile,
+  Link,
+  Navbar,
+  Tile,
+  Header,
+  Footer,
+};
