+++ src/declarations.d.ts
@@ -3,6 +3,19 @@ declare module '*.scss' {
   export = content;
 }
 
+declare module 'react-masonry-css' {
+  import * as React from 'react';
+
+  export interface MasonryProps {
+    breakpointCols?: Object;
+    columnClassName?: string;
+  }
+
+  export default class Masonry extends React.Component<MasonryProps & React.HTMLProps<HTMLElement>, any> {
+    render(): JSX.Element;
+  }
+}
+
 interface PortfolioItem {
   title: string;
   slug: string;
