+++ src/styles/global.scss
@@ -43,7 +43,6 @@ ol, ul {
   // padding-left: 0;
 }
 
-
 .global-brand {
   display: inline-block;
   height: 4rem;
