--- src/templates/react-masonry-css.d.ts
@@ -1,12 +1,0 @@
-declare module 'react-masonry-css' {
-  import * as React from 'react';
-
-  export interface MasonryProps {
-    breakpointCols?: Object;
-    columnClassName?: string;
-  }
-
-  export default class Masonry extends React.Component<MasonryProps & React.HTMLProps<HTMLElement>, any> {
-    render(): JSX.Element;
-  }
-}
